Add the download action to the each list item's action menu. (#43)
* Add the download action to the each list item's action menu.
Adjust layout to accommodate the dropdown menu.

* Make necessary boundary calculations for an above boundary for dropdown menu.
Make primary-content a fixed width for now. This is to make it so when the width is small, the list items don't become jumbled up. Also using word-break: break-word for the title.

* Add a comment about working around a github issue involving dropdown menus inside list items.
Add the overflow-x class to the primary component of the two column layout so the current min-width can be scrolled.