US131040: Track captions and metadata changes separately
When the user attempts to switch locales, Producer will display a
warning dialog if there are unsaved captions changes. This dialog should
not be shown if there are metadata changes, but no captions changes.

If the user clicks "Yes" on this warning dialog and discards the
captions changes to load a different language, the expected behaviour
would be:
- Metadata changes - the "Unsaved" indicator stays "Unsaved"
- No metadata changes - "Unsaved" changes back to "Saved"
Tracking metadata and captions changes separately allows us to keep
the saved/unsaved indicator accurate in this way.