fix: Display captions cues timestamps as non-editable text
Previously, we displayed captions cues' start and end timestamps
in an input text control. This was misleading, as editing the
input text did not actually modify the cue's timestamp data.

In the future, we may come back and allow the user to manually edit the
timestamps using an input text. However, this functionality is not
necessary for release, as it was not implemented in legacy Producer nor
the Audio/Video Note Captions Editor.