fix: Prevent end time from going past file duration when syncing start
When a cue's "Sync start time" button is clicked, and the Media Player's
current time position is later than the cue's end time, the cue's end
time is automatically updated using this formula:
    newEndTime = newStartTime + (oldEndTime - oldStartTime)
to ensure that the cue's duration stays the same.

However, this can cause the new end time to go past the audio/video
file's duration.

The fix is to set end time to the file duration, if the calculated
end time would have gone past it.