Fix 'Edit Live Event' page breaking after first time it is viewed
The 'Edit Live Event' page observes the query parameters in the routing
store. When they change, it then checks if the subView (the last part of
the URL) is set to 'edit' before reloading the page content.

The issue was that the query parameters in the routing store were being
updated BEFORE the subView.

So when the user goes to a different page, the query parameters become
empty and subView is set to undefined. Since the 'Edit Live Event' page
still exists in the background, it updates and tries to fetch with
empty query parameters, resulting in a broken page.

Then when the user goes to the 'Edit Live Event' page again, the
query parameters update. But because this happens before the subView
updates in the routing store, the page does not reload and we see the
broken page as mentioned above.

The solution is simply to update the routing store's subView value
before updating the query parameters, to ensure the page always reloads
as expected.