fix: Fix landing page always redirecting to 404
Originally, the landing page checked the value of the "Can Manage
Capture Central" variable. If false, it would redirect to 404 to prevent
user access. The previous code used an incorrect variable reference,
hence the landing page always redirected to 404.

The proper permission to use in this case would be "Can Access Capture
Central", so this commit adds a getter for it in PermissionStore and
updates the landing page to use it.