fix: Fix d2l-content-uploader not displaying preview when SD is missing
Because of the recent cost-cutting changes to transcoding, a video file
might not have an SD transcode available.

After uploading, d2l-content-uploader's Preview page tries to load both
SD and HD formats. If SD doesn't exist, the 404 causes the dialog to
break.

To fix this, we now check for 404 when fetching each media format. If a
404 is received, then that format isn't passed to the Media Player.

In the exceptional case where all formats return 404, meaning no
transcodes were produced, we display a "This media file failed to be
processed" error.