fix: Prevent wide videos from cutting off controls in Producer
Because the Media Player's dimensions are now controlled by the
"aspect-ratio" CSS property, we need to set an explicit size boundary
for the Media Player in Producer. Otherwise, it will grow too large and
push the tab controls outside of Producer's renderable area.

This commit puts the Media Player inside a container div whose width is
limited to 67% of Producer's overall width. To make this change visually
identical to the existing UI, the div's background color is set to black
so that it resembles pillar boxes on the top and bottom of the player.