fix: Prevent Media Player from unpausing when clicking timeline
When clicking on the timeline, the Media Player pauses to allow the
user to click-and-drag to a specific point. When the mouse is released,
the timeline should make the Media Player unpause if it was previously
playing, or keep it paused if it was previously paused.

To determine whether to pause or unpause, Producer passes down a
'mediaPlayerPaused' attribute to the timeline so that the timeline is
always aware of the Media Player's current state.

There was a bug where 'mediaPlayerPaused' was not being updated
correctly. To ensure that this attribute is always synchronized with the
actual Media Player state, this commit adds listeners for the Media
Player's @pause and @play events that update 'mediaPlayerPaused'
accordingly.