fix: Restrict Media Library Preview width to prevent rapid resizing
(US143731)

This also fixes the Media Player's sizing in the preview dialog so that
it always fits in the dialog's base dimensions, preventing scrollbars
from appearing.

Unfortunately, this does NOT work in desktop Safari -- the existing
behaviour with scrollbars is still present there. But there is no
rapid resizing, at least.